Исправил ошибки. Оставил классы необходимые только в рамках ДЗ16
