--- src/main/java/com/sbrf/reboot/atm/Banknote.java
@@ -1,7 +1,0 @@
-package com.sbrf.reboot.atm;
-
-/**
- * Представляет сущность банкноты
- */
-public class Banknote {
-}
