--- src/main/java/com/sbrf/reboot/atm/Cassette.java
@@ -1,22 +1,0 @@
-package com.sbrf.reboot.atm;
-
-import lombok.AllArgsConstructor;
-
-import java.util.ArrayList;
-
-/**
- * Представляет сущность кассеты с банкнотами
- *
- * @param <T> Может содержать банкноты одного номинала
- */
-@AllArgsConstructor
-public class Cassette<T extends Banknote> {
-    private ArrayList<T> banknotes;
-
-    /**
-     * @return количество банкнот в кассете одного номинала
-     */
-    public int getCountBanknotes() {
-        return banknotes.size();
-    }
-}
