--- src/main/java/com/sbrf/reboot/dto/Request.java
@@ -1,10 +1,0 @@
-package com.sbrf.reboot.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class Request {
-    private final String atmNumber;
-}
