--- src/main/java/com/sbrf/reboot/dto/Response.java
@@ -1,11 +1,0 @@
-package com.sbrf.reboot.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class Response {
-    private final String statusCode;
-
-}
