--- src/main/java/com/sbrf/reboot/repository/AccountRepository.java
@@ -1,10 +1,0 @@
-package com.sbrf.reboot.repository;
-
-import com.sbrf.reboot.dto.Account;
-
-import java.io.IOException;
-import java.util.Set;
-
-public interface AccountRepository {
-    Set<Account> getAllAccountsByClientId(long clientId) throws IOException;
-}
