--- src/main/resources/Accounts.txt
@@ -1,22 +1,0 @@
-[
-  {
-    "clientId": 1,
-    "number": "2-ACCNUM"
-  },
-  {
-    "clientId": 2,
-    "number": "5-ACC1NUM"
-  },
-  {
-    "clientId": 3,
-    "number": "3-ACCNUM"
-  },
-  {
-    "clientId": 1,
-    "number": "X-XXXXXXX"
-  },
-  {
-    "clientId": 1,
-    "number": "4-ACC1NUM"
-  }
-]
